+++ dhall-json/src/Dhall/JSON.hs
@@ -624,12 +624,12 @@ parseOmission :: Parser (Value -> Value)
 parseOmission =
         Options.Applicative.flag'
             omitNull
-            (   Options.Applicative.long "omitNull"
+            (   Options.Applicative.long "omit-null"
             <>  Options.Applicative.help "Omit record fields that are null"
             )
     <|> Options.Applicative.flag'
             omitEmpty
-            (   Options.Applicative.long "omitEmpty"
+            (   Options.Applicative.long "omit-empty"
             <>  Options.Applicative.help "Omit record fields that are null or empty records"
             )
     <|> pure id
@@ -640,7 +640,7 @@ parseNullPreservation =
         Options.Applicative.flag
             omitNull
             id
-            (   Options.Applicative.long "preserveNull"
+            (   Options.Applicative.long "preserve-null"
             <>  Options.Applicative.help "Preserve record fields that are null"
             )
 
@@ -1038,7 +1038,7 @@ parseConversion =
     noConversion =
         Options.Applicative.flag'
             NoConversion
-            (   Options.Applicative.long "noMaps"
+            (   Options.Applicative.long "no-maps"
             <>  Options.Applicative.help "Disable conversion of association lists to homogeneous maps"
             )